refactor(InputChecker) : 입력받은 값에 대한 체크클래스를 따로 외부로추출
