+++ src/test/kotlin/baseball/BaseballModelTest.kt
@@ -1,10 +1,9 @@
 package baseball
 
-import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-
-import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.assertThrows
 
 internal class BaseballModelTest {
 
@@ -15,15 +14,16 @@ internal class BaseballModelTest {
         baseballModel = BaseballModel()
     }
 
+    @Nested
     inner class UserInputTest {
         @Test
         fun `3자리 숫자가 아닌 값이 입력됬을 때_IllegalArgumentException을 발생시킨다`() {
             val testInputs = listOf("1234", "53", "1", "4123123123512312312312313122312312312312312312314123123", "")
             testInputs.forEach { case ->
-                org.junit.jupiter.api.assertThrows<IllegalArgumentException>(
+                assertThrows<IllegalArgumentException>(
                     "$case 3자리가 아니거나, 숫자가 아닌 값"
                 ) {
-                    baseballModel.checkInputIsCorrect(case)
+                    baseballModel.checkBasballNumber(case)
                 }
             }
         }
@@ -32,10 +32,10 @@ internal class BaseballModelTest {
         fun `숫자가 아닌 값이 입력됬을 때_IllegalArgumentException을 발생시킨다`() {
             val testInputs = listOf("asd", "d221sf", "간12", "", " ", "@!2")
             testInputs.forEach { case ->
-                org.junit.jupiter.api.assertThrows<IllegalArgumentException>(
+                assertThrows<IllegalArgumentException>(
                     "$case 3자리가 아니거나, 숫자가 아닌 값"
                 ) {
-                    baseballModel.checkInputIsCorrect(case)
+                    baseballModel.checkBasballNumber(case)
                 }
             }
         }
@@ -44,13 +44,24 @@ internal class BaseballModelTest {
         fun `각각 다른 3자리 숫자가 아닐 때_IllegalArgumentException을 발생시킨다`() {
             val testInputs = listOf("121", "333", "454", "666")
             testInputs.forEach { case ->
-                org.junit.jupiter.api.assertThrows<IllegalArgumentException>(
+                assertThrows<IllegalArgumentException>(
                     "$case 각각 다른 3자리 숫자가 아님"
                 ) {
-                    baseballModel.checkInputIsCorrect(case)
+                    baseballModel.checkBasballNumber(case)
+                }
+            }
+        }
+
+        @Test
+        fun `끝났을 떄의 인풋 테스트`() {
+            val testInputs = listOf("3", "432", "", "aasd")
+            testInputs.forEach { case ->
+                assertThrows<IllegalArgumentException>() {
+                    baseballModel.checkEndedNumber(case)
                 }
             }
         }
     }
+
 
 }
